Fix API startup with robust imports, error handling, and dependency updates
