+++ apps/api/requirements.txt
@@ -6,3 +6,22 @@ joblib>=1.2.0
 pandas>=1.5.0
 numpy>=1.21.0
 scikit-learn>=1.1.0
+
+# MLflow for experiment tracking
+mlflow>=1.30.0
+
+# Database connectors
+psycopg2-binary>=2.9.0
+sqlalchemy>=1.4.0
+
+# Kafka streaming
+kafka-python>=2.0.2
+
+# Additional utilities
+python-dotenv>=0.19.0
+werkzeug>=2.0.0
+
+# Data processing
+scipy>=1.9.0
+matplotlib>=3.5.0
+seaborn>=0.11.0
